From inline leftscreen class into index css
