+++ src/index.css
@@ -2,3 +2,24 @@
 @tailwind components;
 @tailwind utilities;  
   
+.screen {
+    @apply  w-full 
+    flex 
+    items-center 
+    justify-center 
+    h-full 
+    bg-[url('../../public/svg/gradientbg.svg')] 
+    bg-no-repeat 
+    bg-cover 
+} 
+
+.logo {
+    @apply 
+    w-full 
+    h-1/2 
+    bg-cover 
+    bg-center
+   
+    
+}
+
